requirements.txt
#https://download.pytorch.org/whl/cpu/torch-1.5.0%2Bcpu-cp38-cp38-linux_x86_64.whl
torch==1.5.0
torchvision==0.6.1
requests_toolbelt
google-cloud-storage